rewrite

+++ src/甲级/1125.cpp
@@ -12,6 +12,6 @@ int main() {
     int ans = nums[0];
     for (int i = 1; i < n; i++)
         ans = (ans + nums[i]) / 2;
-    cout << (int)ans << endl;
+    cout << ans << endl;
     return 0;
 }
